Extract functions into file
