+++ src/server/actions/Agencies.ts
@@ -10,15 +10,14 @@ import {
   AgencyModel,
   ServiceModel,
 } from '@/server/models';
-import { authOptions } from '@/app/api/auth/[...nextauth]/options';
-import { getServerSession } from 'next-auth';
+import { authenticateServerAction } from '@/utils/auth';
 
 /**
  * @brief Gets all agencies
  * @returns An array of all agencies in the "agencies" collection with fields populated
  */
 export async function getAgencies(): Promise<Agency[]> {
-  authenticateUser();
+  authenticateServerAction();
   try {
     await dbConnect();
     const agencies = await AgencyModel.find({})
@@ -47,7 +46,7 @@ export async function getPaginatedAgencies(
   page: number,
   pageSize: number
 ): Promise<Agency[]> {
-  authenticateUser();
+  authenticateServerAction();
   if (page < 1) {
     throw new JSendResponse({
       status: 'fail',
@@ -116,7 +115,7 @@ export async function getAgencyById(id: string): Promise<Agency> {
  * @throws See mongoErrorHandler for common insertion errors
  */
 export async function createService(service: Service): Promise<Service> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
   const newService = await ServiceModel.create(service).catch((err) => {
     mongoErrorHandler(err);
@@ -131,7 +130,7 @@ export async function createService(service: Service): Promise<Service> {
  * @throws See mongoErrorHandler for common insertion errors
  */
 export async function createAgency(agency: Agency): Promise<Agency> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
 
   const newAgency = await AgencyModel.create(agency).catch((error) => {
@@ -149,7 +148,7 @@ export async function createAgency(agency: Agency): Promise<Agency> {
 export async function createAgencyInfo(
   agencyInfo: AgencyInfoForm
 ): Promise<AgencyInfoForm> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
 
   const newAgencyInfo = await AgencyInfoFormModel.create(agencyInfo).catch(
@@ -171,7 +170,7 @@ export async function updateAgency(
   id: string,
   updates: Partial<Agency>
 ): Promise<Agency | null> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
   const updatedAgency = await AgencyModel.updateOne({ _id: id }, updates).catch(
     (error) => {
@@ -204,7 +203,7 @@ export async function updateService(
   id: string,
   updates: Partial<Service>
 ): Promise<Service | null> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
   const updatedService = await ServiceModel.updateOne(
     { _id: id },
@@ -229,7 +228,7 @@ export async function updateService(
  * @throws 404 if no agency is found with the ID
  */
 export async function deleteAgency(id: string): Promise<Agency | null> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
   const deletedAgency = await AgencyModel.findByIdAndDelete(id).catch(
     (error) => {
@@ -252,7 +251,7 @@ export async function deleteAgency(id: string): Promise<Agency | null> {
  * @throws 404 if no service is found with the ID
  */
 export async function deleteService(id: string): Promise<Service | null> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
   const deletedService = await ServiceModel.findByIdAndDelete(id).catch(
     (error) => {
@@ -281,7 +280,7 @@ export async function approveAgency(
   newInfo: AgencyInfoForm,
   newApprovalStatus: 'Pending' | 'Approved'
 ): Promise<Agency | null> {
-  authenticateUser();
+  authenticateServerAction();
   await dbConnect();
 
   const { services, ...infoWithoutServices } = newInfo;
@@ -368,16 +367,5 @@ function mongoErrorHandler(error: MongoError) {
       // Catch non-mongoose errors
       throw new JSendResponse({ status: 'error', message: errors.serverError });
       break;
-  }
-}
-
-function authenticateUser() {
-  const session = getServerSession(authOptions);
-  if (!session) {
-    throw new JSendResponse({
-      status: 'error',
-      message: 'Unauthorized',
-      code: 401,
-    });
   }
 }
