+++ src/utils/auth.ts
@@ -1,0 +1,27 @@
+import { authOptions } from '@/app/api/auth/[...nextauth]/options';
+import { getServerSession } from 'next-auth';
+import { JSendResponse } from './types';
+
+export function authenticateServerAction() {
+  const session = getServerSession(authOptions);
+  if (!session) {
+    throw new JSendResponse({
+      status: 'error',
+      message: 'Unauthorized',
+      code: 401,
+    });
+  }
+}
+
+export async function authenticateServerEndpoint() {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return Response.json(
+      new JSendResponse({
+        status: 'error',
+        message: 'Unauthorized',
+      }),
+      { status: 401 }
+    );
+  }
+}
